pyproject.toml
[tool.poetry]
name = "worknet"
version = "0.1.0"
description = ""
authors = ["Kait <qw.kaitlyn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
pytz = "^2024.1"
django-htmx = "^1.19.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
social-auth-app-django = "^5.4.2"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
pre-commit = "^3.8.0"
django-debug-toolbar = "^4.4.6"
commitizen = "^3.29.0"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
